+++ yandex/resource_yandex_organizationmanager_group_iam_member.go
@@ -1,0 +1,7 @@
+package yandex
+
+import "github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+
+func resourceYandexOrganizationManagerGroupIAMMember() *schema.Resource {
+	return resourceIamMemberWithImport(IamGroupSchema, newGroupIamUpdater, groupIDParseFunc)
+}